Fixed: All forms were being bound but could never submit because _checkFormSubmission always returned false;
I rearranged the way 404's are thrown so that runRoute always raises 404s.
If silence_404 is on you will never see them unless you call runRoute() directly.
Instead _checkURL and _checkFormSubmission catch the 404 and decide what to do.
This is actually a good improvement as it means you never have to return from post() routes,
but if you want to you could do

  post('#/post-form', function() { with(this) {
    // do something to the form
    if (something) { // if something goes wrong or isnt correct
      return true; // will actually submit the form to the server
    }
  }});