Support for query strings in the URL fragment.
Adds support for key-value parameters in a query string that appears
after the hash.  The query string is ignored when looking up routes; so
you can pass arbitrary parameters to a Sammy action without changing the
route.

Query string parameters are mapped to the `params` object just
like path parameters and form values.  For example, this link:

    <a href="#/docs?page=3&per_page=15">next page of results</a>

Will route to this Sammy action:

    app = $.sammy(function() {
        this.get('#/docs', function() {
            show_results_page(this.params.page, this.params.per_page);
        });
    });

The syntax for query strings in the URL fragment is the same for
traditional query strings: the query is separated from the path by a
question mark; parameters are separated by ampersands; keys and values
are separated by equals signs.