Fix for IE6 bug involving question mark in hash.
If a question mark appears in a URL fragment, IE6 will exclude the
question mark and everything after it from `window.location.hash`.  For
example, given the location:

    http://example.com/#/docs?page=1&per_page=10

The value of `window.location.hash` will be:

    #/docs

By bypassing `window.location.hash` and parsing the URL within Sammy to
extract the fragment we can avoid cross-browser inconsistencies like
this one.