Rebuilt the internal event system to use jQuery's built in namespacing.
This means a couple things:
Applications can now catch normal events fired within the element that the app is bound to. However, within Application#bind() callbacks `this` is a Sammy.EventContext, even if the events were fired by elements and not with app#trigger().

Also, app#trigger formerly took an optional third argument which was the context in which to evaluate the callback. Instead, you can pass {context: your_desired_context} as the data argument to trigger.