New partial system dont care about your rendering engine.
Partial now figures out the rendering 'engine' based on the final extension of the path your fetching. An 'engine' is just a method that takes a string and an object. See the Sammy.Template plugin as an example.

Also if you include Sammy.Cache it will cache your requested files unless you say cache_partials = false.