Sammy() is a function itself (instead of just an object).
Sammy() is the top level object that is not a constructor, but a looker-uper
Sammy.apps holds an object that maps a selector to a single application. This
allows you to lookup and easily fetch/extend an app object from anywhere.
It also prevents multiple apps for selectors, which becomes hairy when trying to
figure out which app is handling which event. However, if you want to have multiple
apps on single selectors, its still possible - just use the new Sammy.Application()
(constructor) syntax.