partial() will now iterate over an array of data.
if the data param passed to partial() is an Array, after fetching the template,
partial will iterate over the data, concatenating the result, or calling the
callback for each iteration. No matter what (even if Sammy.Cache isn't included)
it will only fetch the partial from the server once.

Example:

    // name.template
    <li><%= name %></li>

    // in sammy app
    this.get('#/names', function() {
      this.partial('name.template', [{name: 'AQ', name: 'MB'}])
    });

    // element now looks like:
    <li>AQ</li><li>MB</li>