Fixed issue with DataLocationProxy#getLocation() returning null
Changed DataLocationProxy to work like HashChangeProxy so that it
returns an empty string rather than null when the current location
has not been set.  This fixes the issue that causes the app below to
fail with a 'path is null' error.

var app = $.sammy('div.app',function(a){
  a.setLocationProxy(new Sammy.DataLocationProxy(this));
  a.get('#/',function(){
    this.swap('Hello world')
  })
});
app.run('#/');